Move readwrite splitting statement to readwrite splitting module
